+++ lib/cypressTagsHelpers.test.js
@@ -1,4 +1,4 @@
-const { stripCLIArguments, getGlobArg } = require("./tagsHelper");
+const { stripCLIArguments, getGlobArg } = require("./cypressTagsHelpers");
 
 const processArgv = process.argv;
